+++ CHANGES.md
@@ -40,6 +40,11 @@
 - [FIX] mid を nullable に変更する
     - 「type: offer の mid を必須にする」の対応で role が recvonly の時にエラーとなる不具合の修正
     - @miosakuma
+- [FIX] EGLContext が取れなかった場合、DefaultVideoDecoderFactory, SoraDefaultVideoEncoderFactory を使用する
+    - EGLContext が取れなかった場合の Decoder を SoftwareVideoDecoderFactory から DefaultVideoDecoderFactory に変更する
+    - EGLContext が取れなかった場合の Encoder を SoftwareVideoEncoderFactory から SoraDefaultVideoEncoderFactory に変更する
+    - EGLContext は null でも Hardware を使用する MediaCodec は動作するため HW も動作可能な DefaultVideoDecoderFactory, SoraDefaultVideoEncoderFactory を使用する
+    - @miosakuma
 
 ## 2022.3.0